Moved UrlGenerator::Generators to separate files
